--- distribution/pull_v2_unix.go
@@ -1,13 +1,0 @@
-// +build !windows
-
-package distribution
-
-import (
-	"github.com/docker/distribution"
-	"github.com/docker/distribution/context"
-)
-
-func (ld *v2LayerDescriptor) open(ctx context.Context) (distribution.ReadSeekCloser, error) {
-	blobs := ld.repo.Blobs(ctx)
-	return blobs.Open(ctx, ld.digest)
-}
