--- layer/layer_store_windows.go
@@ -1,11 +1,0 @@
-package layer
-
-import (
-	"io"
-
-	"github.com/docker/distribution"
-)
-
-func (ls *layerStore) RegisterWithDescriptor(ts io.Reader, parent ChainID, descriptor distribution.Descriptor) (Layer, error) {
-	return ls.registerWithDescriptor(ts, parent, descriptor)
-}
